refactor: reorganize User and News modules, update namespaces, and remove unused models
