+++ app/Modules/User/Controllers/UserController.php
@@ -6,7 +6,7 @@ use App\Http\Controllers\BaseController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
-use App\Models\User;
+use App\Modules\User\Model\User;
 
 class UserController extends BaseController
 {
