+++ app/Providers/AppServiceProvider.php
@@ -4,6 +4,7 @@ namespace App\Providers;
 
 use Illuminate\Support\Facades\View;
 use Illuminate\Support\ServiceProvider;
+use InvalidArgumentException;
 
 class AppServiceProvider extends ServiceProvider
 {
@@ -12,7 +13,7 @@ class AppServiceProvider extends ServiceProvider
      */
     public function register(): void
     {
-        //
+
     }
 
     /**
@@ -27,18 +28,23 @@ class AppServiceProvider extends ServiceProvider
                 continue;
             }
 
-            $routeFile = $modulesPath . '/' . $module . '/config/routes.php';
+            $routeFile = $modulesPath . '/' . $module . '/Config/routes.php';
 
             $viewPath = $modulesPath . '/' . $module . '/Views';
             
             if (is_dir($viewPath)) {
-                
+
                 View::addNamespace(strtolower($module), $viewPath);
             }
 
-            if (is_file($routeFile)) {
+            if (file_exists($routeFile)) {
                 require $routeFile;
+
+            } else {
+                
+                throw new InvalidArgumentException("Failed to resolve model class for the route file: {$routeFile}. Please check the file structure and namespace definitions.");
             }
+            
         }
     }
